pyproject.toml
[project]
name = "learning-roadmap-generation"
version = "0.1.0"
description = "An AI-powered agent designed to assist users in creating personalized learning roadmaps."
authors = [
    {name = "Paoladevelopment",email = "paolala1017@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9.4,<4.0"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "openai (>=1.66.3,<2.0.0)",
    "langchain-openai (>=0.3.9,<0.4.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (>=4.11.2,<5.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "litellm (>=1.63.14,<2.0.0)",
    "langgraph (>=0.3.18,<0.4.0)",
    "langchain-tavily (>=0.1.5,<0.2.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "langgraph-cli[inmem] (>=0.1.89,<0.2.0)",
]

[tool.poetry]
packages = [{include = "learning_roadmap_generation", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
